Had to externalize TokenEnchant's event listener for the TEBlockExplodeEvent so it won't cause any errors when TE is not installed.  It was preventing AutoFeatures from loading when it goes wrong.
