+++ prison-spigot/src/main/java/tech/mcprison/prison/spigot/autofeatures/AutoManagerFeatures.java
@@ -29,6 +29,7 @@ import tech.mcprison.prison.Prison;
 import tech.mcprison.prison.autofeatures.AutoFeaturesFileConfig.AutoFeatures;
 import tech.mcprison.prison.internal.block.PrisonBlock;
 import tech.mcprison.prison.mines.data.Mine;
+import tech.mcprison.prison.mines.features.MineBlockEvent.BlockEventType;
 import tech.mcprison.prison.output.Output;
 import tech.mcprison.prison.spigot.SpigotPrison;
 import tech.mcprison.prison.spigot.SpigotUtil;
@@ -107,6 +108,188 @@ public class AutoManagerFeatures
 		return (short) itemInHand.getBukkitStack().getEnchantmentLevel(Enchantment.LOOT_BONUS_BLOCKS);
 	}
 
+	
+    
+    
+    @Override
+	public boolean doAction( SpigotBlock block, Mine mine, Player player ) {
+    	
+    	return applyAutoEvents( block, player, mine );
+	}
+    
+    
+    /**
+     * <p>This function overrides the doAction in OnBlockBreakEventListener and
+     * this is only enabled when auto manager is enabled.
+     * </p>
+     * 
+     */
+    @Override
+    public boolean doAction( Mine mine, Player player, List<SpigotBlock> explodedBlocks, 
+    								BlockEventType blockEventType, String triggered ) {
+    	return applyAutoEvents( player, mine, explodedBlocks, blockEventType, triggered );
+    }
+	
+
+	private boolean applyAutoEvents( SpigotBlock spigotBlock, Player player, Mine mine) {
+		boolean cancel = false;
+		
+		if (isBoolean(AutoFeatures.isAutoManagerEnabled) && 
+			!spigotBlock.isEmpty() ) {
+			
+//			Output.get().logInfo( "#### AutoManager.applyAutoEvents: BlockBreakEvent: :: " + mine.getName() + "  " + 
+//					"  blocks remaining= " + 
+//					mine.getRemainingBlockCount() + " [" + block.toString() + "]"
+//					);
+
+			SpigotItemStack itemInHand = SpigotPrison.getInstance().getCompatibility().getPrisonItemInMainHand( player );
+
+			int count = applyAutoEvents( player, spigotBlock, mine );
+			
+			if ( count > 0 ) {
+				processBlockBreakage( spigotBlock, mine, player, count, BlockEventType.blockBreak,
+										null, itemInHand );
+
+    			cancel = true;
+			}
+			
+			if ( mine != null ) {
+				checkZeroBlockReset( mine );
+			}
+	
+		}
+		
+		return cancel;
+	}
+
+
+	
+	
+	private int applyAutoEvents( Player player, SpigotBlock block, Mine mine ) {
+		int count = 0;
+		
+		SpigotItemStack itemInHand = SpigotPrison.getInstance().getCompatibility().getPrisonItemInMainHand( player );
+
+		
+		boolean isLoreEnabled = isBoolean( AutoFeatures.isLoreEnabled );
+		
+		boolean lorePickup = isLoreEnabled && checkLore( itemInHand, getMessage( AutoFeatures.lorePickupValue ) );
+		boolean loreSmelt = isLoreEnabled && checkLore( itemInHand, getMessage( AutoFeatures.loreSmeltValue) );
+		boolean loreBlock = isLoreEnabled && checkLore( itemInHand, getMessage( AutoFeatures.loreBlockValue ) );
+		
+		
+		boolean isAutoPickup = lorePickup || isBoolean( AutoFeatures.autoPickupEnabled ) ||
+										player.isPermissionSet( getMessage( AutoFeatures.permissionAutoPickup ));
+		
+		boolean isAutoSmelt = loreSmelt || isBoolean( AutoFeatures.autoSmeltEnabled ) ||
+										player.isPermissionSet( getMessage( AutoFeatures.permissionAutoPickup ));
+		
+		boolean isAutoBlock = loreBlock || isBoolean( AutoFeatures.autoBlockEnabled ) ||
+										player.isPermissionSet( getMessage( AutoFeatures.permissionAutoBlock ));
+		
+		// NOTE: Using isPermissionSet so players that are op'd to not auto enable everything.
+		//       Ops will have to have the perms set to actually use them.
+				
+		// AutoPickup
+		if ( (mine != null || mine == null && !isBoolean( AutoFeatures.autoPickupLimitToMines )) &&
+				isAutoPickup ) {
+			
+			count = autoFeaturePickup( block, player, itemInHand );
+
+			// Cannot set to air yet, or auto smelt and auto block will only get AIR:
+//			autoPickupCleanup( block, count );
+		}
+		
+		XMaterial source = SpigotUtil.getXMaterial( block.getPrisonBlock() );
+		
+		// AutoSmelt
+		if ( (mine != null || mine == null && !isBoolean( AutoFeatures.autoSmeltLimitToMines )) &&
+				isAutoSmelt ){
+			
+			// Smelting needs to change the source to the smelted target so auto block will work:
+			source = autoFeatureSmelt( player, source );
+		}
+		
+		// AutoBlock
+		if ( (mine != null || mine == null && !isBoolean( AutoFeatures.autoBlockLimitToMines )) &&
+				isAutoBlock ) {
+			
+			autoFeatureBlock( player, source );
+		}
+		
+		
+		// AutoPickup - Clean up (set block to air)
+		if ( (mine != null || mine == null && !isBoolean( AutoFeatures.autoPickupLimitToMines )) &&
+				isAutoPickup ) {
+			
+			autoPickupCleanup( block, count );
+			
+		}
+
+		
+		return count;
+	}
+
+
+	/**
+	 * <p>This function gets called for EACH block that is impacted by the
+	 * explosion event.  The event may have have a list of blocks, but not all
+	 * blocks may be included in the mine. This function is called ONLY when
+	 * a block is within a mine,
+	 * </p>
+	 * 
+	 * <p>The event TEBlockExplodeEvent has already taken place and it handles all
+	 * actions such as auto pickup, auto smelt, and auto block.  The only thing we
+	 * need to do is to record the number of blocks that were removed during 
+	 * the explosion event. The blocks should have been replaced by air.
+	 * </p>
+	 * 
+	 * <p>Normally, prison based auto features, and or the perms need to be enabled 
+	 * for the auto features to be enabled, but since this is originating from another
+	 * plugin and another external configuration, we cannot test for any real perms
+	 * or configs.  The fact that this event happens within one of the mines is 
+	 * good enough, and must be counted.
+	 * </p>
+	 * 
+	 * @param e
+	 * @param mine
+	 */
+	private boolean applyAutoEvents( Player player, Mine mine, 
+									List<SpigotBlock> explodedBlocks, BlockEventType blockEventType, 
+										String triggered ) {
+		boolean cancel = false;
+		
+		int totalCount = 0;
+
+		SpigotItemStack itemInHand = SpigotPrison.getInstance().getCompatibility().getPrisonItemInMainHand( player );
+
+		
+		// The explodedBlocks list have already been validated as being within the mine:
+		for ( SpigotBlock spigotBlock : explodedBlocks ) {
+			
+			int drop = applyAutoEvents( player, spigotBlock, mine );
+			totalCount += drop;
+			
+			if ( drop > 0 ) {
+				
+				processBlockBreakage( spigotBlock, mine, player, drop, 
+											blockEventType, triggered, itemInHand );
+			}
+		}
+		
+		if ( mine != null ) {
+			checkZeroBlockReset( mine );
+		}
+		
+		if ( totalCount > 0 ) {
+			cancel = true;
+		}
+		
+		return cancel;
+	}
+	
+	
+	
 	/*
 	 * Drops are wrong with old 1.14.4 releases of spigot
 	 * but got fixed in newer versions.
